added test and fix code smells
